backend/requirements.txt
# Requirements for Hotel Management System Backend
# Install using: pip install -r requirements.txt

Flask==2.3.0
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.0
openpyxl==3.1.2
reportlab==4.0.4
